+++ replace_windows.go
@@ -1,0 +1,41 @@
+//go:build windows
+
+package main
+
+import (
+	"syscall"
+	"unsafe"
+)
+
+var moveFileExW = syscall.NewLazyDLL("kernel32.dll").NewProc("MoveFileExW")
+
+func atomicReplace(source, destination string) error {
+	from, err := syscall.UTF16PtrFromString(source)
+	if err != nil {
+		return err
+	}
+	to, err := syscall.UTF16PtrFromString(destination)
+	if err != nil {
+		return err
+	}
+	const (
+		moveFileReplaceExisting = 0x1
+		moveFileWriteThrough    = 0x8
+	)
+	result, _, callErr := moveFileExW.Call(
+		uintptr(unsafe.Pointer(from)),
+		uintptr(unsafe.Pointer(to)),
+		uintptr(moveFileReplaceExisting|moveFileWriteThrough),
+	)
+	if result != 0 {
+		return nil
+	}
+	if callErr != syscall.Errno(0) {
+		return callErr
+	}
+	return syscall.EINVAL
+}
+
+func syncDirectory(string) error {
+	return nil
+}